Create Non-ascii tokenization
Tokenize all the words in german_text using word_tokenize(), and print the result.
Tokenize only the capital words in german_text.
First, write a pattern called capital_words to match only capital words. Make sure to check for the German Ü! To use this character in the exercise, copy and paste it from these instructions.
Then, tokenize it using regexp_tokenize().
Tokenize only the emoji in german_text. The pattern using the unicode ranges for emoji given in the assignment text has been written for you. Your job is to use regexp_tokenize() to tokenize the emoji.