Create Charting practice
Use the .split() method on holy_grail with the newline character ('\n') as the argument.
Recall that re.sub() requires 3 arguments: The pattern, the replacement, and the string. The pattern is given for you; the replacement is '' and the string is l.
Use regexp_tokenize() as the output expression of your list comprehension, with s and "\w+" as the arguments.
To create line_num_words, use len(t_line) as the output expression of the list comprehension.
Use plt.hist() with line_num_words as the argument to create the histogram, and then plt.show() to display it.